third_party/WebKit/LayoutTests/compositing/squashing/dont-squash-into-animated-layers.html

<!doctype HTML>
Any errors will show below this line.
<div id="target" style="position: absolute; width: 200px; height: 200px; will-change: transform; transition: transform 0.1s cubic-bezier(0.23, 1, 0.32, 1); background: lightblue"></div>
<div style="position: absolute; width: 200px; height: 200px; top: 100px; left: 100px; background: lightgray"></div>
<script src="../../resources/testharness.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
onload = function() {
    target.style.transform = "translateX(10px)";
    requestAnimationFrame(function() {
        if (window.internals) {
            var layers = JSON.parse(internals.layerTreeAsText(document, 1));
            assert_true(layers.children[0].children[1].compositingReasons[1] == "Cannot squash into a layer that is animating.");
        }
        testRunner.notifyDone();
    });
};
</script>
